Decouple recording metrics
